+++ Packages/com.unity.cloud.gltfast.tests/Tests/Runtime/Scripts/Import/DataUriTests.cs
@@ -15,6 +15,10 @@ namespace GLTFast.Tests.Import
     {
         const string k_TestDataBase64 = "rQbwDQ=="; // AD06F00D, a dog food ;)
         const string k_TestDataBase64Invalid = "rQbw}Q==";
+        /// <summary>One black pixel WebP</summary>
+        const string k_TestWebP = "UklGRiQAAABXRUJQVlA4IBgAAAAwAQCdASoBAAEAAgA0JaQAA3AA/vv9UAA=";
+        const string k_TestKtxInvalid = "q0tUWCAyMLsNChoK";
+        const string k_TestKtx1Px = "q0tUWCAyMLsNChoKKwAAAAEAAAABAAAAAQAAAAAAAAAAAAAAAQAAAAEAAAAAAAAAaAAAAFwAAADEAAAAMAAAAAAAAAAAAAAAAAAAAAAAAAD0AAAAAAAAAAQAAAAAAAAABAAAAAAAAABcAAAAAAAAAAIAWAABAQIAAAAAAAQAAAAAAAAAAAAHAAAAAAAAAAAA/wAAAAgABwEAAAAAAAAAAP8AAAAQAAcCAAAAAAAAAAD/AAAAGAAHHwAAAAAAAAAA/wAAACwAAABLVFh3cml0ZXIAa3R4IGNyZWF0ZSB2NC40LjAgLyBsaWJrdHggdjQuNC4wAAAAAP8=";
 
         [UnityTest]
         public IEnumerator BufferDataUriUnexpectedMimeType()
@@ -67,16 +71,16 @@ namespace GLTFast.Tests.Import
         [UnityTest]
         public IEnumerator ImageDataUriUnexpectedMimeType()
         {
-            var gltf = $@"{{""images"":[{{""uri"":""data:image/fantasy-format;base64,{k_TestDataBase64Invalid}""}}],""textures"":[{{""source"":0}}]}}";
-            var task = Test(gltf, true, LogCode.EmbedImageLoadFailed);
+            var gltf = $@"{{""images"":[{{""uri"":""data:image/fantasy-format;base64,{k_TestDataBase64}""}}],""textures"":[{{""source"":0}}]}}";
+            var task = Test(gltf, true, LogCode.ImageFormatUnknown);
             yield return AsyncWrapper.WaitForTask(task);
         }
 
         [UnityTest]
         public IEnumerator ImageDataUriUnsupportedMimeType()
         {
-            var gltf = $@"{{""images"":[{{""uri"":""data:image/webp;base64,{k_TestDataBase64Invalid}""}}],""textures"":[{{""source"":0}}]}}";
-            var task = Test(gltf, true, LogCode.EmbedImageUnsupportedType);
+            var gltf = $@"{{""images"":[{{""uri"":""data:image/webp;base64,{k_TestWebP}""}}],""textures"":[{{""source"":0}}]}}";
+            var task = Test(gltf, true, LogCode.ImageFormatUnsupported);
             yield return AsyncWrapper.WaitForTask(task);
         }
 
@@ -94,11 +98,7 @@ namespace GLTFast.Tests.Import
             var gltf = $@"{{""images"":[{{""uri"":""data:image/ktx2;base64,{k_TestDataBase64Invalid}""}}],""textures"":[{{""source"":0}}]}}";
             var task = Test(
                 gltf, true,
-#if KTX_IS_RECENT
                 LogCode.EmbedImageLoadFailed
-#else
-                LogCode.EmbedImageUnsupportedType
-#endif
                 );
             yield return AsyncWrapper.WaitForTask(task);
         }
@@ -106,27 +106,44 @@ namespace GLTFast.Tests.Import
         [UnityTest]
         public IEnumerator ImageKtxContentBroken()
         {
-            var gltf = $@"{{""images"":[{{""uri"":""data:image/ktx2;base64,{k_TestDataBase64}""}}],""textures"":[{{""source"":0}}]}}";
+#if KTX_IS_RECENT
+            LogAssert.Expect(LogType.Error, "KTX error code FileUnexpectedEof");
+#endif
+            var gltf = $@"{{""images"":[{{""uri"":""data:image/ktx2;base64,{k_TestKtxInvalid}""}}],""textures"":[{{""source"":0}}]}}";
             var task = Test(
                 gltf, true,
 #if KTX_IS_RECENT
                 LogCode.EmbedImageLoadFailed
 #else
-                LogCode.EmbedImageUnsupportedType
+                LogCode.PackageMissing
 #endif
                 );
             yield return AsyncWrapper.WaitForTask(task);
+        }
+
+        [UnityTest]
+        public IEnumerator ImageKtxNoMipMaps()
+        {
 #if KTX_IS_RECENT
-            LogAssert.Expect(LogType.Error, "KTX error code FileUnexpectedEof");
+            LogAssert.Expect(LogType.Warning, "KTX texture does not contain mipmaps.");
 #endif
+            var gltf = $@"{{""images"":[{{""uri"":""data:image/ktx2;base64,{k_TestKtx1Px}""}}],""textures"":[{{""source"":0}}]}}";
+            var task = Test(
+                gltf, true
+#if !KTX_IS_RECENT
+                , LogCode.PackageMissing
+#endif
+                );
+            yield return AsyncWrapper.WaitForTask(task);
         }
 
         static async Task Test(string gltf, bool expectSuccess, params LogCode[] expectedLogCodes)
         {
             var logger = new CollectingLogger();
             var import = new GltfImport(logger: logger);
-            Assert.AreEqual(expectSuccess, await import.LoadGltfJson(gltf));
-            Assert.AreEqual(2, logger.Count);
+            var settings = new ImportSettings { GenerateMipMaps = true };
+            Assert.AreEqual(expectSuccess, await import.LoadGltfJson(gltf, importSettings: settings));
+            Assert.AreEqual(1 + expectedLogCodes.Length, logger.Count);
             LoggerTest.AssertLogger(logger, expectedLogCodes);
         }
     }